+++ NEWS.md
@@ -136,6 +136,7 @@ Red Panda C++ Version 2.27
   - enhancement: In debug console, Ctrl+C/Ctrl+X/Ctrl+V conflicts with application action.
   - enhancement: Auto hide Edit/Selection/Code/Refactor menu if no file openning.
   - enhancement: Auto hide Project menu if no project openning.
+  - fix: Toggle breakpoint by shortcut may use wrong line.
     
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.